chore(commands): One file per command PE-342
