+++ src/commands/create_drive.ts
@@ -3,11 +3,8 @@ import { ArFSDAO } from '../arfsdao';
 import { CLICommand } from '../CLICommand';
 import { CommonContext } from '../CLICommand/common_context';
 import {
-	DriveIdParameter,
-	DriveKeyParameter,
 	DriveNameParameter,
 	DrivePasswordParameter,
-	GetAllRevisionsParameter,
 	SeedPhraseParameter,
 	WalletFileParameter
 } from '../parameter_declarations';
@@ -32,27 +29,6 @@ new CLICommand({
 		})();
 		console.log(JSON.stringify(createDriveResult, null, 4));
 
-		process.exit(0);
-	}
-});
-
-new CLICommand({
-	name: 'drive-info',
-	parameters: [
-		DriveIdParameter,
-		GetAllRevisionsParameter,
-		DrivePasswordParameter,
-		DriveKeyParameter,
-		WalletFileParameter
-	],
-	async action(options) {
-		const context = new CommonContext(options);
-		const wallet = await context.getWallet();
-		const arDrive = new ArDrive(new ArFSDAO(wallet, arweave));
-		const driveId: string = options.driveId;
-		// const getAllRevisions: boolean = options.getAllRevisions;
-		const result = await arDrive.getPublicDrive(driveId /*, getAllRevisions*/);
-		console.log(JSON.stringify(result, null, 4));
 		process.exit(0);
 	}
 });
